rudimentary implementation of file backing
For sparse matrices, this currently assumes that the indices are sorted.
This appears to be the case for my test file, but is not given in
general, since AnnData does not sort indices prior to writing them to
h5ad. Supporting general sparse matrices would require reimplementing
getindex methods from scratch