string indexing for AnnData
This is currently very slow. If adata has N names and we want to extract
K names, the complexity is O(NK). Ideally we would use an ordered set
for row_names and var_names, but we need the ordered set to support both
key and index lookup as well as being able to return an index for a
given key. OrderedSet from OrderedCollections.jl deprecated index lookup
and does not have an API for looking up an index for a given key (see
https://github.com/JuliaCollections/OrderedCollections.jl/issues/64 and
https://github.com/JuliaCollections/DataStructures.jl/pull/180#issuecomment-400269803