MuData: correct subsetting in the presence of duplicates (see #4)
The previous approach to subset involved finding the MuData rownames and
searching for those rownames in the individual AnnData objects.
Duplicates were resolved based on the position of the corresponding rows
in the MuData object. This approach can fail when duplicates are
present, since the join operation responsible for generating the MuData
names can and probably will completely reorder the data frame, therefore
the position in the MuData object is uncorrelated with the the position
in the individual AnnData objects.

The new approach maintains a mapping between MuData positions and
AnnData positions. This leads to much simpler subsetting code that
generates unambiguous results.