fix obs/varmaps in subset objects with non-continous indices
This now does the same as Python muon. Note that in contrast to muon,
varmaps in views still refere to the original objects and cannot be used
for subsetting from the view. This doesn't affect us, since we always
subset from the parent object, and leads to faster and more light-weight
views, but might be confusing to users. However, these are not primarily
meant for users anyway, so...